requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
